+++ LP50.py
@@ -15,6 +15,7 @@ class LP50(base.Component):
     """
     # RELEASES
     VERSION = base.VersionCollection(
+        base.VersionInfo("2.2.10", "2023-09-20"),
         base.VersionInfo("2.2.9", "2023-09-19"),
         base.VersionInfo("2.2.8", "2023-09-18"),
         base.VersionInfo("2.2.7", "2023-09-13"),
@@ -81,6 +82,8 @@ class LP50(base.Component):
     VERSION.changed("2.2.8", "Updated input descriptions and removed stub descriptions")
     VERSION.added("2.2.8", "Runtime note regarding removal of SimulationStart input")
     VERSION.changed("2.2.9", "Extended description of LP50 output")
+    VERSION.changed("2.2.10", "Extended output descriptions")
+    VERSION.changed("2.2.10", "`LP50` output reports reach geometries")
 
     def __init__(self, name, observer, store):
         """
@@ -151,7 +154,10 @@ class LP50(base.Component):
                     "shape": (
                         "the same years as the [Values](#Values) input",
                         "the same reaches as the [Values](#Values) input"
-                    )
+                    ),
+                    "element_names": (None, "according to the `Values` input"),
+                    "offset": ("the year of the `SimulationStart` input", None),
+                    "geometries": (None, "according to the `Values` input")
                 }
             )
         ])
@@ -171,10 +177,11 @@ class LP50(base.Component):
         """
         processing_path = self._inputs["ProcessingPath"].read().values
         reaches = self.inputs["Values"].describe()["element_names"][1]
+        geometries = self.inputs["Values"].describe()["geometries"][1]
         simulation_start = self._inputs["SimulationStart"].read().values
         self.prepare_module_inputs(processing_path, reaches.get_values(), simulation_start)
         self.run_module(processing_path)
-        self.read_module_outputs(processing_path, reaches, simulation_start)
+        self.read_module_outputs(processing_path, reaches, simulation_start, geometries)
         return
 
     def prepare_module_inputs(self, processing_path, reaches, simulation_start):
@@ -225,7 +232,13 @@ class LP50(base.Component):
         )
         return
 
-    def read_module_outputs(self, processing_path, reaches, simulation_start):
+    def read_module_outputs(
+            self,
+            processing_path: str,
+            reaches: np.ndarray,
+            simulation_start: datetime.date,
+            geometries: list[bytes]
+    ) -> None:
         """
         Reads the module's outputs and stores them in the Landscape model store.
 
@@ -233,6 +246,7 @@ class LP50(base.Component):
             processing_path: The working directory of the module.
             reaches: The identifiers of the reaches considered.
             simulation_start: The first day of the simulation.
+            geometries: The geometries of the reaches in Well-Known-Byte notation.
 
         Returns:
             Nothing.
@@ -255,5 +269,5 @@ class LP50(base.Component):
             elif value == -999:
                 value = error_report_value
             lp50[time_index, space_index] = value
-        self._outputs["LP50"].set_values(lp50, element_names=(None, reaches), offset=(simulation_start.year, None))
-        return
+        self._outputs["LP50"].set_values(
+            lp50, element_names=(None, reaches), offset=(simulation_start.year, None), geometries=(None, geometries))